+++ l10n_br_base/views/res_company_view.xml
