+++ l10n_br_base/views/res_country_view.xml
