+++ l10n_br_base/views/res_partner_view.xml
